+++ src/render.ts
@@ -1,4 +1,10 @@
-import { Page } from './types';
+export type Page = {
+  body: string;
+  suffix: string;
+  description: string;
+  title: string;
+  archiveHeader: string;
+};
 
 const rootElement = <HTMLDivElement>document.getElementById('root');
 const archiveHeaderElement = <HTMLHeadingElement>(
