--- src/types.ts
@@ -1,7 +1,0 @@
-export type Page = {
-  body: string;
-  suffix: string;
-  description: string;
-  title: string;
-  archiveHeader: string;
-};
